그룹/스팟 개설 신청 MapStruct 작성 및 로그인 로직 복구

+++ app-public-api/src/main/java/co/kurrant/app/public_api/controller/client/ApplicationFormController.java
@@ -61,7 +61,7 @@ public class ApplicationFormController {
 
     @Operation(summary = "기업 스팟 개설 신청 내역", description = "기업 스팟 개설 신청 내역을 조회한다.")
     @GetMapping("/corporations/{id}")
-    public ResponseMessage getCorporationApplicationFormDetail(HttpServletRequest httpServletRequest, @PathVariable Long id) {
+    public ResponseMessage getCorporationApplicationFormDetail(HttpServletRequest httpServletRequest, @PathVariable BigInteger id) {
         return ResponseMessage.builder()
                 .message("기업 스팟 신청 내역 조회에 성공하였습니다.")
                 .data(applicationFormService.getCorporationApplicationFormDetail(commonService.getUserId(httpServletRequest), id))
@@ -70,7 +70,7 @@ public class ApplicationFormController {
 
     @Operation(summary = "기업 스팟 개설 신청 내역 기타 내용 저장", description = "기업 스팟 개설 신청 내역 기타 내용을 저장한다.")
     @PutMapping("/corporations/{id}/memo")
-    public ResponseMessage SaveCorporationsApplicationFormMemo(HttpServletRequest httpServletRequest, @PathVariable Long id, @RequestBody ApplicationFormMemoDto applicationFormMemoDto) {
+    public ResponseMessage SaveCorporationsApplicationFormMemo(HttpServletRequest httpServletRequest, @PathVariable BigInteger id, @RequestBody ApplicationFormMemoDto applicationFormMemoDto) {
         applicationFormService.updateCorporationApplicationFormMemo(httpServletRequest, id, applicationFormMemoDto);
         return ResponseMessage.builder()
                 .message("기업 스팟 개설 신청 내역 기타 내용 수정에 성공하였습니다.")
