+++ domain-application-form/src/main/java/co/dalicious/domain/application_form/repository/CorporationApplicationFormRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.math.BigInteger;
 import java.util.List;
 
-public interface CorporationApplicationFormRepository extends JpaRepository<CorporationApplicationForm, Long> {
+public interface CorporationApplicationFormRepository extends JpaRepository<CorporationApplicationForm, BigInteger> {
     List<CorporationApplicationForm> findAllByUserId(BigInteger id);
 }
